Separate schemas for notes and versions
