+++ src/schemas/notes.py
@@ -3,14 +3,7 @@ from typing import List, Optional
 
 from pydantic import BaseModel, Field
 
-
-class BaseModelPaginated(BaseModel):
-    """Base schema for paginated responses."""
-
-    prev_page: Optional[str] = Field(None, description="URL for previous page")
-    next_page: Optional[str] = Field(None, description="URL for next page")
-    total_pages: int = Field(..., description="Total number of pages")
-    total_items: int = Field(..., description="Total number of notes")
+from schemas import VersionDetailResponseSchema
 
 
 class NoteDetailResponseSchema(BaseModel):
@@ -21,8 +14,13 @@ class NoteDetailResponseSchema(BaseModel):
     versions: List["VersionDetailResponseSchema"]
 
 
-class NoteListResponseSchema(BaseModelPaginated):
+class NoteListResponseSchema(BaseModel):
     notes: List[NoteDetailResponseSchema]
+
+    prev_page: Optional[str] = Field(None, description="URL for previous page")
+    next_page: Optional[str] = Field(None, description="URL for next page")
+    total_pages: int = Field(..., description="Total number of pages")
+    total_items: int = Field(..., description="Total number of notes")
 
 
 class NoteCreateRequestSchema(BaseModel):
@@ -31,15 +29,3 @@ class NoteCreateRequestSchema(BaseModel):
 
 class NoteUpdateRequestSchema(BaseModel):
     content: str = Field(..., description="The updated content of the note")
-
-
-class VersionDetailResponseSchema(BaseModel):
-    id: int
-    note_id: int
-    version: int
-    content: str
-    created_at: datetime
-
-
-class VersionListResponseSchema(BaseModelPaginated):
-    versions: List[VersionDetailResponseSchema]
